Extract debug_print to a separate header file.
